+++ src/index.ts
@@ -1,1 +1,2 @@
-export { getChatCompletion } from './inference';
+export * as openAi from './providers/openAi';
+export * as perplexity from './providers/perplexity';
